resolve ui with interface
