--- appcompose/src/main/java/com/picone/appcompose/ui/component/screen/HomeEventManager.kt
@@ -1,20 +1,0 @@
-package com.picone.appcompose.ui.component.screen
-
-import androidx.compose.runtime.Composable
-import com.picone.appcompose.ui.component.manager.action.Action
-import com.picone.appcompose.ui.component.screen.home.HomeActions
-import com.picone.appcompose.ui.component.screen.home.HomeTaskEvents
-
-class HomeEventManager {
-
-   /* val stateManager : HomeStateManager = HomeStateManager()
-
-    @Composable
-    fun triggerEvent(action : Action){
-        when(action){
-            HomeActions.updateScreenState-> stateManager.updateState(HomeTaskEvents.GET_INITIAL_STATE)
-        }
-    }*/
-
-
-}
