--- appcompose/src/main/java/com/picone/appcompose/ui/component/screen/detail/DetailActions.kt
@@ -1,4 +1,0 @@
-package com.picone.appcompose.ui.component.screen.detail
-
-object DetailActions {
-}
