--- appcompose/src/main/java/com/picone/appcompose/ui/component/screen/home/HomeActions.kt
@@ -1,14 +1,0 @@
-package com.picone.appcompose.ui.component.screen.home
-
-import com.picone.appcompose.ui.component.manager.action.Action
-
-object HomeActions {
-
-    val updateScreenState = object:Action{}
-    val topAppBarAddItemButtonOpenPoUp = object : Action {}
-    val topAppBarAddItemButtonClosePoUp = object : Action {}
-    val navigateToAddOnPopUpItemSelected = object : Action {}
-    val navigateToDetailOnTaskClicked = object : Action{}
-    val navigateToHomeTaskOnNavItemClicked = object : Action{}
-    val navigateToProjectOnNavItemClicked = object : Action{}
-}
