--- appcompose/src/main/java/com/picone/appcompose/ui/component/screen/home/HomeEvents.kt
@@ -1,10 +1,0 @@
-package com.picone.appcompose.ui.component.screen.home
-
-import androidx.compose.runtime.Composable
-import androidx.lifecycle.ViewModel
-import com.picone.appcompose.ui.component.screen.Event
-
-object HomeEvents {
-
-    val onScreenCreated = object : Event{}
-}
