+++ appcompose/src/main/java/com/picone/appcompose/ui/main/screen/detail/DetailActionManager.kt
@@ -1,0 +1,4 @@
+package com.picone.appcompose.ui.main.screen.detail
+
+class DetailActionManager {
+}
