+++ appcompose/src/main/java/com/picone/appcompose/ui/main/screen/detail/DetailActions.kt
@@ -1,0 +1,4 @@
+package com.picone.appcompose.ui.main.screen.detail
+
+object DetailActions {
+}
